Keep agent harness boundary tests within reviewable modules
The red and green work made the harness contract file too large. Splitting fixture construction and the three-agent boundary cases keeps the test surface small without changing the verifier behavior.

Constraint: issue #7054 follows the mandatory refactor phase and the repo line-budget guidance

Rejected: leave the oversized test files until later | the issue can be kept small now

Confidence: high

Scope-risk: narrow

Directive: Keep future harness boundary cases in the focused submodules instead of growing the root integration test file

Tested: CARGO_TARGET_DIR=target/mvp-demo-proof CARGO_BUILD_JOBS=1 CARGO_INCREMENTAL=0 cargo test -p kamn-e2e-harness --test mvp_demo_agent_harness_claim_contract -- --nocapture

Tested: cargo fmt --check
