+++ crates/kamn-e2e-harness/tests/mvp_demo_agent_harness_claim_contract/support.rs
@@ -2,7 +2,10 @@ use std::path::{Path, PathBuf};
 
 use kamn_e2e_harness::{MvpDemoCommandConfig, VerifyMvpDemoCommandConfig};
 
+mod artifact;
 mod three_agent;
+
+pub(crate) use artifact::*;
 
 pub(crate) fn temp_root(stem: &str) -> PathBuf {
     let millis = std::time::SystemTime::now()
@@ -80,71 +83,6 @@ pub(crate) fn report_with_three_agent_claim(root: &Path, artifact: &Path) -> Str
     )
 }
 
-pub(crate) fn agent_artifact(root: &Path, private_visible: bool, settlement_label: &str) -> String {
-    agent_artifact_with_surface(root, private_visible, settlement_label, "mcp-tools")
-}
-
-pub(crate) fn agent_artifact_without_three_agent_boundary(root: &Path) -> String {
-    agent_artifact_for_report_with_surface(
-        root.join("proof/report.json")
-            .display()
-            .to_string()
-            .as_str(),
-        false,
-        "devnet-backed",
-        "pi-extension-tools",
-        three_agent::NO_THREE_AGENT_BOUNDARY,
-    )
-}
-
-pub(crate) fn agent_artifact_with_surface(
-    root: &Path,
-    private_visible: bool,
-    settlement_label: &str,
-    execution_surface: &str,
-) -> String {
-    agent_artifact_for_report_with_surface(
-        root.join("proof/report.json")
-            .display()
-            .to_string()
-            .as_str(),
-        private_visible,
-        settlement_label,
-        execution_surface,
-        three_agent::absent_boundary(),
-    )
-}
-
-pub(crate) fn agent_latest_artifact(root: &Path) -> String {
-    agent_latest_artifact_with_surface(root, "mcp-tools")
-}
-
-pub(crate) fn agent_latest_artifact_with_surface(root: &Path, execution_surface: &str) -> String {
-    agent_artifact_for_report_with_surface(
-        root.join("latest/proof/report.json")
-            .display()
-            .to_string()
-            .as_str(),
-        false,
-        "devnet-backed",
-        execution_surface,
-        three_agent::absent_boundary(),
-    )
-}
-
-pub(crate) fn agent_artifact_with_three_agent_boundary(root: &Path) -> String {
-    agent_artifact_for_report_with_surface(
-        root.join("proof/report.json")
-            .display()
-            .to_string()
-            .as_str(),
-        false,
-        "devnet-backed",
-        "pi-extension-tools",
-        three_agent::valid_boundary(),
-    )
-}
-
 fn write_file(path: &Path, content: String) {
     std::fs::create_dir_all(path.parent().expect("parent should exist"))
         .expect("fixture parent directory should be creatable");
@@ -203,17 +141,4 @@ fn agent_claim() -> &'static str {
 
 fn roadmap_claim() -> &'static str {
     r#"{"id":"production_readiness","label":"roadmap","required":false,"status":"NOT_CLAIMED","summary":"production readiness is not claimed"}"#
-}
-
-fn agent_artifact_for_report_with_surface(
-    report_path: &str,
-    private_visible: bool,
-    settlement_label: &str,
-    execution_surface: &str,
-    three_agent_boundary: &str,
-) -> String {
-    format!(
-        r#"{{"schema_version":"kamn.mvp.agent-harness-evidence.v1","harness":"mcp-agent","execution_surface":"{}","report_path":"{}","verifier_status":"PASS","participant_agents":["agent_a","agent_b","agent_c_verifier"],"tool_markers":["register","create_task","fund_escrow","release_escrow","verify_proof"],"claim_boundaries":{{"settlement_claim_label":"{}","dry_run_counted_as_success":false,"placeholder_counted_as_success":false,"verifier_private_view_visible":{}}}{}}}"#,
-        execution_surface, report_path, settlement_label, private_visible, three_agent_boundary
-    )
 }
